# docs/api/v0.openapi.yaml
openapi: 3.1.0
info:
  title: Fishhawk Backend API
  version: 0.1.0
  description: |
    The HTTP surface that the Fishhawk backend (`fishhawkd`) exposes to its
    three classes of caller:

    1. **The Web UI and CLI** — humans driving plan review, approval, audit
       search, and run management.
    2. **The runner** — `kuhlman-labs/fishhawk/runner@vX.Y` running on the
       customer's CI, requesting per-run signing keys and shipping signed
       trace bundles.
    3. **GitHub webhooks** — the customer's GitHub App installation pushing
       issue / PR / check events that drive workflow dispatch.

    Per ADR (#46) and `MVP_SPEC.md` §5.1.1, every customer-facing endpoint is
    prefixed with `/v0`. The `v0` prefix is **unstable** by design — the
    surface is allowed to change before v1. Once the v1 surface is published
    (post-Day 21), `/v0` and `/v1` coexist for one minor version cycle to
    let clients migrate, and `/v0` is then retired.

    Webhook ingestion (`/webhooks/github`) is unversioned because GitHub
    controls the request shape; the backend adapts as GitHub's webhook
    schema evolves.

  contact:
    name: Fishhawk
    url: https://github.com/kuhlman-labs/fishhawk
  license:
    name: Apache-2.0
    url: https://github.com/kuhlman-labs/fishhawk/blob/main/LICENSE
servers:
  - url: https://api.fishhawk.dev
    description: Production
  - url: http://localhost:8080
    description: Local dev (docker compose up)

...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: fishhawk_session
      description: |
        HTTP-only `Secure; SameSite=Lax` session cookie set by the OAuth
        callback. Same-origin SPA requests are auto-included; CSRF
        protection on state-changing endpoints (POST/PUT/PATCH/DELETE)
        requires an `X-CSRF-Token` request header whose value matches
        the `__Host-csrf` cookie issued alongside the session cookie.
        Mismatched / missing token → `403 csrf_required`. Bearer-token
        clients (see `bearerToken`) are immune to CSRF and bypass the
        header check.

    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Scoped server-issued token (NOT a JWT). Index into the
        `api_tokens` table; revocation is immediate. Used by the CLI
        and any programmatic caller.

    githubOIDC:
      type: http
      scheme: bearer
      bearerFormat: jwt
      description: |
        GitHub Actions OIDC token for runner authentication. Verified
        by the backend against GitHub's JWKS per request; no server-
        side state. Required by `POST /v0/runs/{run_id}/signing-key`.

  # -------------------------------------------------------------------------
  # Shared schemas. Detailed property schemas grow as endpoints are
  # implemented — what's frozen here is the envelope shape and the
  # canonical types every endpoint returns.
  # -------------------------------------------------------------------------
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: |
                Stable, machine-readable code. Snake-case identifier;
                clients switch on this. New codes are additive; codes
                are never repurposed.
              example: run_not_found
            message:
              type: string
              description: |
                Human-readable description. May change between
                versions; clients should not parse it.
              example: "No run with id 11111111-…"
            details:
              type: object
              additionalProperties: true
              description: |
                Optional code-specific structured fields. The keys
                under `details` are part of the contract for that
                specific code.
      examples:
        - error:
            code: run_not_found
            message: No run with id 11111111-2222-3333-4444-555555555555
...
    UUID:
      type: string
      format: uuid
...
    Stage:
      type: object
      required:
        [id, run_id, sequence, type, executor, state, created_at,
         updated_at]
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        run_id: { $ref: '#/components/schemas/UUID' }
        sequence:
          type: integer
          minimum: 0
        type:
          type: string
          enum: [plan, implement, review]
        executor:
          type: object
          description: Exactly one of agent or human.
          properties:
            kind:
              type: string
              enum: [agent, human]
            ref:
              type: string
              example: claude-code
        state:
          type: string
          enum:
            [pending, dispatched, running, awaiting_approval, succeeded,
             failed, cancelled]
        started_at:
          type: [string, "null"]
          format: date-time
        ended_at:
          type: [string, "null"]
          format: date-time
        failure_category:
          type: [string, "null"]
          enum: [A, B, C, D, null]
          description: |
            Set on failed transitions per MVP_SPEC §6.
            A: agent failure, B: constraint/policy, C: infrastructure,
            D: approval timeout.
        failure_reason:
          type: [string, "null"]
        gate:
          oneOf:
            - type: "null"
            - $ref: '#/components/schemas/StageGate'
          description: |
            The persisted shape of the stage's workflow-spec gate.
            Null/omitted when the stage has no gate. Per migration
            0014 / #213 — the review-stage UI reads this to decide
            whether to surface the approval panel. Pre-#254 the
            shape carried a `blocking_checks` array; v0.2 dropped
            that field (ADR-017 / #249) — required CI checks now
            live in branch protection, snapshotted onto the run
            row's `required_checks_snapshot` (#251).
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    StageGate:
      type: object
      description: |
        Workflow-spec gate snapshot stored on the stage row. Mirrors
        spec.Gate — `type` distinguishes approval-vs-check, approvers
        is set only for approval gates. The pre-#254 `blocking_checks`
        array was removed in v0.2 (ADR-017 / #249); required CI
        checks now live in branch protection, exposed via the run's
        `required_checks_snapshot` (#251) and the per-stage
        `/v0/stages/{id}/checks` endpoint.
      required: [type]
      properties:
        type:
          type: string
          enum: [approval, check]
        approvers:
          oneOf:
            - type: "null"
            - type: object
              description: |
                Roles whose members can satisfy the gate. Exactly
                one of any_of / all_of is set when populated.
              properties:
                any_of:
                  type: array
                  items: { type: string }
                all_of:
                  type: array
                  items: { type: string }
...
  responses:
...
    Unauthorized:
      description: Authentication missing or invalid.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: |
        Authenticated identity lacks permission for the operation
        (`error: forbidden`), or — on state-changing methods from a
        session-cookie identity — the `X-CSRF-Token` header didn't
        match the `__Host-csrf` cookie (`error: csrf_required`).
        Bearer-token clients bypass the CSRF check.
        Bearer tokens that authenticate successfully but lack a required
        write scope receive `error: insufficient_scope` with HTTP 403.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Referenced resource does not exist.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
security:
  - sessionCookie: []
  - bearerToken: []
paths:
...
  /v0/auth/github/login:
    get:
      operationId: githubLogin
      summary: Begin GitHub OAuth sign-in
      description: |
        Redirects the browser to GitHub's OAuth authorize endpoint with a
        backend-generated state value. The state is bound to the browser
        via a temporary cookie and verified on callback.
      security: []
      tags: [auth]
      responses:
        '302':
          description: Redirect to GitHub's OAuth authorize URL.
        '500':
          description: OAuth state could not be issued.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /v0/stages/{stage_id}:
    get:
      operationId: getStage
      summary: Get a stage
      tags: [stages]
      parameters:
        - { name: stage_id, in: path, required: true, schema: { $ref: '#/components/schemas/UUID' } }
      responses:
        '200':
          description: Stage details.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Stage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /v0/stages/{stage_id}/retry:
    post:
      operationId: retryStage
      summary: Retry a failed stage
      description: |
        Re-opens a failed stage when the failure category supports it.
        Per MVP_SPEC §6:

        - **D, sla_timeout**: transitions failed → awaiting_approval,
          clears failure metadata, restarts the SLA clock implicitly
          via the `updated_at` trigger. Returns 200 with the updated stage.
        - **A** / **C**: transitions failed → pending, then the
          orchestrator walks pending → dispatched and fires
          workflow_dispatch. Returns 200 with the post-orchestrator
          stage (typically `dispatched`). Orchestrator failures are
          logged but don't fail the request — the stage stays at
          pending and an operator can re-fire Advance manually.
        - **D, gate-rejected**: 422 (`retry_not_applicable`) — the
          approver said no; create a fresh run instead.
        - **B**: 422 (`retry_not_applicable`) — the workflow or spec
          needs to change before a retry would behave differently.

        Writes a `stage_retried` audit entry on success with the prior
        category and reason in the payload.
      tags: [stages]
      parameters:
        - { name: stage_id, in: path, required: true, schema: { $ref: '#/components/schemas/UUID' } }
      responses:
        '200':
          description: Stage re-opened; refreshed Stage row in body.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Stage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422':
          description: |
            Retry not applicable for the stage's current state /
            failure category. Body's `error` is `retry_not_applicable`
            and `message` explains the specific reason.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...